+++ Sources/CardPayments/CardClientError.swift
@@ -39,7 +39,7 @@ enum CardClientError {
         case malformedDeeplinkURLError
 
         /// 10. Cancellation from 3DS verification
-        case threeDSCancellation
+        case threeDSCancellationError
     }
 
     static let unknownError = CoreSDKError(
@@ -68,10 +68,10 @@ enum CardClientError {
         errorDescription: "An invalid 3DS URL was returned. Contact developer.paypal.com/support."
     )
 
-    static let threeDSecureCancellation = CoreSDKError(
-        code: Code.threeDSCancellation.rawValue,
+    static let threeDSecureCanceled = CoreSDKError(
+        code: Code.threeDSCancellationError.rawValue,
         domain: domain,
-        errorDescription: "3DS verification has been cancelled by the user."
+        errorDescription: "3DS verification has been canceled by the user."
     )
 
     static let noVaultTokenDataError = CoreSDKError(